fix getting address of cluster 0
Due to the changes in 530b663 to suuport larger disk the math to covert a
to an address underflows then it cast to a wider int without sign extention
leaving a bit set in the middle of the number.

This change moves the subtraction to happen after the other additions
and widening thus prevting any underflow issues.